refactor: split integrations/github/helpers.py by role
- availability.py: github_source_available()
- credentials.py: GITHUB_INJECTED_PARAMS + github_creds()
- mcp_config.py: resolve_github_mcp_config() + private helper
- result_normalizer.py: normalize_github_tool_result() + private helper
- Delete helpers.py grab-bag
- Update all 16 import sites across tools/ and tests/

Fixes #5370
