--- integrations/github/helpers.py
@@ -1,171 +1,0 @@
-"""Helpers shared by GitHub MCP tools.
-
-Each function takes inputs produced by the rest of the runtime
-(integration store entries, MCP tool results, runtime-extracted kwargs)
-and returns the shape the next layer expects. Callers live in
-``integrations/github/tools/*.py`` and rely on these wrappers so the
-tool files stay thin.
-
-Exports:
-
-- ``GITHUB_INJECTED_PARAMS``: kwargs ``extract_params`` may inject that must
-  win over model-supplied values at call time.
-- ``github_source_available``: predicate on the integration store entry.
-- ``github_creds``: maps classified integration fields to tool kwargs.
-- ``resolve_github_mcp_config``: merges env defaults with explicit overrides.
-- ``normalize_github_tool_result``: turns a raw MCP tool result into the
-  payload shape consumed by the tool framework.
-"""
-
-from __future__ import annotations
-
-import json
-from typing import Any
-
-from core.tool_framework.utils import tool_unavailable
-from integrations.github.mcp import (
-    DEFAULT_GITHUB_MCP_MODE,
-    GitHubMCPConfig,
-    build_github_mcp_config,
-    github_mcp_config_from_env,
-)
-
-# Runtime connection/secret kwargs from ``extract_params``; must win over model input.
-GITHUB_INJECTED_PARAMS: tuple[str, ...] = (
-    "github_url",
-    "github_mode",
-    "github_token",
-    "github_command",
-    "github_args",
-)
-
-
-def github_source_available(sources: dict[str, dict]) -> bool:
-    """Return True when the GitHub integration is configured and reachable.
-
-    ``sources`` is the per-integration view assembled by the runtime from the
-    integration store; the relevant entry is ``sources["github"]``. Returns
-    True only when that entry's ``connection_verified`` flag is set truthy
-    (typically by the verifier after a live credentials check). Missing
-    ``github`` entry or a falsy/missing ``connection_verified`` returns False.
-    """
-    return bool(sources.get("github", {}).get("connection_verified"))
-
-
-def github_creds(gh: dict) -> dict[str, Any]:
-    """Map classified GitHub integration fields to tool credential kwargs."""
-    creds: dict[str, Any] = {}
-    url = gh.get("github_url") or gh.get("url")
-    if url:
-        creds["github_url"] = url
-    mode = gh.get("github_mode") or gh.get("mode")
-    if mode:
-        creds["github_mode"] = mode
-    token = gh.get("github_token") or gh.get("auth_token")
-    if token:
-        creds["github_token"] = token
-    command = gh.get("github_command") or gh.get("command")
-    if command:
-        creds["github_command"] = command
-    args = gh.get("github_args")
-    if args is None:
-        args = gh.get("args")
-    if args:
-        creds["github_args"] = list(args)
-    return creds
-
-
-def _has_explicit_github_mcp_overrides(
-    github_url: str | None,
-    github_mode: str | None,
-    github_token: str | None,
-    github_command: str | None,
-    github_args: list[str] | None,
-) -> bool:
-    if github_url or github_token or github_command or github_args:
-        return True
-    return bool(github_mode and github_mode != DEFAULT_GITHUB_MCP_MODE)
-
-
-def resolve_github_mcp_config(
-    github_url: str | None,
-    github_mode: str | None,
-    github_token: str | None,
-    github_command: str | None = None,
-    github_args: list[str] | None = None,
-) -> GitHubMCPConfig | None:
-    """Return the GitHub MCP config to use, merging env defaults with overrides.
-
-    Reads ``github_mcp_config_from_env()`` for the env-derived baseline, then
-    treats any non-default value among ``github_url``, ``github_token``,
-    ``github_command``, ``github_args``, or a non-default ``github_mode`` as an
-    explicit override. When no overrides are present, returns the env config
-    as-is. Otherwise builds a fresh ``GitHubMCPConfig`` filling unset fields
-    from the env config (or ``DEFAULT_GITHUB_MCP_MODE`` for ``mode`` when no
-    env value is available) and returns it.
-    """
-    env_config = github_mcp_config_from_env()
-    if not _has_explicit_github_mcp_overrides(
-        github_url, github_mode, github_token, github_command, github_args
-    ):
-        return env_config
-    return build_github_mcp_config(
-        {
-            "url": github_url or (env_config.url if env_config else ""),
-            "mode": github_mode or (env_config.mode if env_config else DEFAULT_GITHUB_MCP_MODE),
-            "auth_token": github_token or (env_config.auth_token if env_config else ""),
-            "command": github_command or (env_config.command if env_config else ""),
-            "args": github_args or (list(env_config.args) if env_config else []),
-            "headers": env_config.headers if env_config else {},
-            "toolsets": env_config.toolsets if env_config else (),
-        }
-    )
-
-
-def _structured_content_or_text_fallback(result: dict[str, Any]) -> Any:
-    """Return ``structured_content`` if present, else ``text`` parsed as JSON.
-
-    Returns ``None`` when ``text`` is empty or not valid JSON.
-    """
-    structured = result.get("structured_content")
-    if structured is not None:
-        return structured
-    text = str(result.get("text") or "").strip()
-    if not text:
-        return None
-    try:
-        return json.loads(text)
-    except json.JSONDecodeError:
-        return None
-
-
-def normalize_github_tool_result(result: dict[str, Any]) -> dict[str, Any]:
-    """Normalize a raw GitHub MCP tool result into the tool-framework payload.
-
-    ``result`` is the dict returned by ``call_github_mcp_tool``: it carries
-    ``is_error`` (bool), ``text`` (str, root-cause message on error),
-    ``tool`` (str, the MCP tool name), ``arguments`` (dict passed to the tool),
-    ``structured_content`` (parsed JSON or None), and ``content`` (list of MCP
-    content items). When ``is_error`` is truthy, returns the standard
-    ``tool_unavailable("github", ...)`` envelope so the framework surfaces a
-    consistent unavailable-source response. Otherwise returns a dict with
-    ``source="github"``, ``available=True``, and the original ``tool``,
-    ``arguments``, ``text``, ``content`` keys preserved, and
-    ``structured_content`` normalized via :func:`_structured_content_or_text_fallback`.
-    """
-    if result.get("is_error"):
-        return tool_unavailable(
-            "github",
-            result.get("text") or "GitHub MCP tool call failed.",
-            tool=result.get("tool"),
-            arguments=result.get("arguments", {}),
-        )
-    return {
-        "source": "github",
-        "available": True,
-        "tool": result.get("tool"),
-        "arguments": result.get("arguments", {}),
-        "text": result.get("text", ""),
-        "structured_content": _structured_content_or_text_fallback(result),
-        "content": result.get("content", []),
-    }
