+++ integrations/github/tools/git_deploy_timeline_tool/__init__.py
@@ -26,14 +26,11 @@ from core.domain.types.incident_window import IncidentWindow
 from core.domain.types.tools import ToolSurface
 from core.tool_framework.tool_decorator import tool
 from core.tool_framework.utils import tool_unavailable
-from integrations.github.helpers import (
-    GITHUB_INJECTED_PARAMS,
-    github_creds,
-    github_source_available,
-    normalize_github_tool_result,
-    resolve_github_mcp_config,
-)
+from integrations.github.availability import github_source_available
+from integrations.github.credentials import GITHUB_INJECTED_PARAMS, github_creds
 from integrations.github.mcp import call_github_mcp_tool
+from integrations.github.mcp_config import resolve_github_mcp_config
+from integrations.github.result_normalizer import normalize_github_tool_result
 
 DEFAULT_WINDOW_MINUTES = 120
 MAX_WINDOW_MINUTES = 7 * 24 * 60  # 7 days
