+++ tests/tools/test_github_file_contents_tool.py
@@ -43,7 +43,7 @@ def test_extract_params_maps_fields() -> None:
 
 
 def test_run_returns_unavailable_when_no_config() -> None:
-    with patch("integrations.github.helpers.github_mcp_config_from_env", return_value=None):
+    with patch("integrations.github.mcp_config.github_mcp_config_from_env", return_value=None):
         result = get_github_file_contents(owner="org", repo="repo", path="README.md")
     assert result == {
         "source": "github",
@@ -64,9 +64,9 @@ def test_run_happy_path() -> None:
     }
     mock_config = MagicMock()
     with (
-        patch("integrations.github.helpers.github_mcp_config_from_env", return_value=None),
+        patch("integrations.github.mcp_config.github_mcp_config_from_env", return_value=None),
         patch(
-            "integrations.github.helpers.build_github_mcp_config",
+            "integrations.github.mcp_config.build_github_mcp_config",
             return_value=mock_config,
         ),
         patch(
@@ -107,9 +107,9 @@ def test_run_falls_back_to_resource_text_when_structured_content_absent() -> Non
     }
     mock_config = MagicMock()
     with (
-        patch("integrations.github.helpers.github_mcp_config_from_env", return_value=None),
+        patch("integrations.github.mcp_config.github_mcp_config_from_env", return_value=None),
         patch(
-            "integrations.github.helpers.build_github_mcp_config",
+            "integrations.github.mcp_config.build_github_mcp_config",
             return_value=mock_config,
         ),
         patch(
