+++ tools/system/python_execution_tool/credentials.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from typing import Any
 
 from integrations.github.client import resolve_github_token
-from integrations.github.helpers import github_creds
+from integrations.github.credentials import github_creds
 
 GITHUB_TOKEN_ENV = "GITHUB_TOKEN"
